Move AbstractJettyHomeTest to jetty-testers module to avoid creating jar and dependency not really needed
